Make SCPRequestPipeline a child class of TxrxProcess
Started to do the cleanup that that allows
