+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/BMPCommandProcess.java
@@ -22,7 +22,7 @@ import static org.slf4j.LoggerFactory.getLogger;
 import static uk.ac.manchester.spinnaker.messages.Constants.BMP_TIMEOUT;
 import static uk.ac.manchester.spinnaker.messages.scp.SequenceNumberSource.SEQUENCE_LENGTH;
 import static uk.ac.manchester.spinnaker.transceiver.ProcessException.makeInstance;
-import static uk.ac.manchester.spinnaker.transceiver.SCPRequestPipeline.RETRY_DELAY_MS;
+import static uk.ac.manchester.spinnaker.transceiver.TxrxProcess.RETRY_DELAY_MS;
 import static uk.ac.manchester.spinnaker.utils.UnitConstants.MSEC_PER_SEC;
 
 import java.io.IOException;
@@ -43,7 +43,6 @@ import uk.ac.manchester.spinnaker.machine.HasCoreLocation;
 import uk.ac.manchester.spinnaker.messages.bmp.BMPRequest;
 import uk.ac.manchester.spinnaker.messages.bmp.BMPRequest.BMPResponse;
 import uk.ac.manchester.spinnaker.messages.scp.SCPResultMessage;
-import uk.ac.manchester.spinnaker.utils.UsedInJavadocOnly;
 import uk.ac.manchester.spinnaker.utils.ValueHolder;
 
 /**
@@ -54,9 +53,7 @@ import uk.ac.manchester.spinnaker.utils.ValueHolder;
  * @param <R>
  *            The type of the response; implicit in the type of the request.
  * @author Donal Fellows
- * @see SCPRequestPipeline
  */
-@UsedInJavadocOnly(SCPRequestPipeline.class)
 class BMPCommandProcess<R extends BMPResponse> {
 	private static final Logger log = getLogger(BMPCommandProcess.class);
